move rotate_zeta to compat
